// src/api/generated/model/apiResponseExamTypeListResponse.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ExamTypeListResponse } from "./examTypeListResponse";

export interface ApiResponseExamTypeListResponse {
  success?: boolean;
  message?: string;
  data?: ExamTypeListResponse;
  errors?: string[];
  timestamp?: string;
  path?: string;
}