// src/api/generated/model/apiResponseExamSessionDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ExamSessionDto } from "./examSessionDto";

export interface ApiResponseExamSessionDto {
  success?: boolean;
  message?: string;
  data?: ExamSessionDto;
  errors?: string[];
  timestamp?: string;
  path?: string;
}